backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
gevent>=1.4
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.5
Flask-Migrate==4.0.5
